removed old scr_chatmode splitscreen can now properly be configured via build configs, and potentially more than just 4 seats. fix splitscreen and enemycolour forcing. forcing now works, but only when all seats are on/spectating the same team. fix demo playback issue with http urls with webgl that don't allow streaming. srgb fixes for the scoreboard.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5229 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/net_wins.c
@@ -6805,7 +6805,7 @@ void NET_PrintConnectionsStatus(ftenet_connections_t *collection)
 int TCP_OpenStream (netadr_t *remoteaddr)
 {
 #ifndef HAVE_TCP
-	return INVALID_SOCKET;
+	return (int)INVALID_SOCKET;
 #else
 	unsigned long _true = true;
 	int newsocket;
